LayoutTests/fast/parser/tag-with-exclamation-point-expected.txt

!foo elements count: 0

!doctype elements count: 0

!foo with id "foo": null

!doctype with id "doctype": null

!foo div child count: 0

!doctype div child count: 0
